+++ climpred/alignment.py
@@ -42,42 +42,24 @@ def return_inits_and_verif_dates(forecast, verif, alignment, reference=None, his
     units = forecast["lead"].attrs["units"]
     leads = forecast["lead"].values
 
-    # `init` renamed to `time` in compute functions.
     all_inits = forecast["init"]
     all_verifs = verif["time"]
 
     # If aligning reference='uninitialized', need to account for potential differences
     # in its temporal coverage. Note that the reference='uninitialized' only aligns
     # verification dates and doesn't care about inits.
-    if hist is not None:
+    if hist is not None and "uninitialzed" not in reference:
         all_verifs = np.sort(list(set(all_verifs.data) & set(hist["time"].data)))
         all_verifs = xr.DataArray(all_verifs, dims=["time"], coords=[all_verifs])
-
-    # print('all_verifs',all_verifs)
-    # print('all_inits',all_inits)
 
     # Construct list of `n` offset over all leads.
     n, freq = get_multiple_lead_cftime_shift_args(units, leads)
 
-    if "time" not in forecast.coords:
-        # print('_construct_init_lead_matrix')
+    if "time" not in forecast.coords:  # construct time(init, lead) in compute_hindcast
         init_lead_matrix = _construct_init_lead_matrix(forecast, n, freq, leads)
         init_lead_matrix["lead"].attrs = forecast.lead.attrs
-        # print(init_lead_matrix)
-    else:
-        # print('dont _construct_init_lead_matrix, use existing')
-        # dim = "time"
-        # not_dim = "init"
-        init_lead_matrix_new = forecast["time"].drop(
-            "time"
-        )  # .drop(not_dim)  # .rename({'init':'time'})
-        # print(init_lead_matrix_new)
-        # init_lead_matrix_new["lead"].attrs = forecast.lead.attrs
-        init_lead_matrix_new = init_lead_matrix_new.rename(None)
-        init_lead_matrix = init_lead_matrix_new
-        # print(init_lead_matrix_new.name,'\n', init_lead_matrix.name)
-    # dim = 'time'
-    # print('ilm',init_lead_matrix.coords, init_lead_matrix.dims)
+    else:  # use time(init, lead)
+        init_lead_matrix = forecast["time"].drop("time").rename(None)
 
     if dask.is_dask_collection(init_lead_matrix):
         init_lead_matrix = init_lead_matrix.compute()
@@ -110,17 +92,16 @@ def _maximize_alignment(init_lead_matrix, all_verifs, leads):
     """
     # Move row-wise and find all forecasted times that align with verification dates at
     # the given lead.
-    dim = "init"  # was time
     verify_with_observations = init_lead_matrix.isin(all_verifs)
     lead_dependent_verif_dates = init_lead_matrix.where(verify_with_observations)
     # Probably a way to do this more efficiently since we're doing essentially
     # the same thing at each step.
     verif_dates = {
-        lead: lead_dependent_verif_dates.sel(lead=lead).dropna(dim).to_index()
+        lead: lead_dependent_verif_dates.sel(lead=lead).dropna("init").to_index()
         for lead in leads
     }
     inits = {
-        lead: lead_dependent_verif_dates.sel(lead=lead).dropna(dim)[dim]
+        lead: lead_dependent_verif_dates.sel(lead=lead).dropna("init")["init"]
         for lead in leads
     }
     return inits, verif_dates
@@ -132,12 +113,11 @@ def _same_inits_alignment(init_lead_matrix, valid_inits, all_verifs, leads, n, f
 
     See ``return_inits_and_verif_dates`` for descriptions of expected variables.
     """
-    dim = "init"  # was time
     verifies_at_all_leads = init_lead_matrix.isin(all_verifs).all("lead")
     inits = valid_inits.where(verifies_at_all_leads, drop=True)
     inits = {lead: inits for lead in leads}
     verif_dates = {
-        lead: shift_cftime_index(inits[lead], dim, n, freq)
+        lead: shift_cftime_index(inits[lead], "init", n, freq)
         for (lead, n) in zip(leads, n)
     }
     return inits, verif_dates
@@ -149,9 +129,8 @@ def _same_verifs_alignment(init_lead_matrix, valid_inits, all_verifs, leads, n, 
 
     See ``return_inits_and_verif_dates`` for descriptions of expected variables.
     """
-    dim = "init"  # was time
     common_set_of_verifs = [
-        i for i in all_verifs if (i == init_lead_matrix).any(dim).all("lead")
+        i for i in all_verifs if (i == init_lead_matrix).any("init").all("lead")
     ]
     if not common_set_of_verifs:
         raise CoordinateError(
@@ -160,7 +139,7 @@ def _same_verifs_alignment(init_lead_matrix, valid_inits, all_verifs, leads, n, 
             "'same_inits' or 'maximize'."
         )
     # Force to CFTimeIndex for consistency with `same_inits`
-    verif_dates = xr.concat(common_set_of_verifs, dim).to_index()
+    verif_dates = xr.concat(common_set_of_verifs, "init").to_index()
     inits_that_verify_with_verif_dates = init_lead_matrix.isin(verif_dates)
     inits = {
         lead: valid_inits.where(
@@ -174,6 +153,8 @@ def _same_verifs_alignment(init_lead_matrix, valid_inits, all_verifs, leads, n, 
 
 def _construct_init_lead_matrix(forecast, n, freq, leads):
     """Returns xr.DataArray of "real time" (init + lead) over all inits and leads.
+    Identical results as add_time_from_init_lead, which also works for non integer n.
+    _construct_init_lead_matrix is only used in compute_hindcast now.
 
     Arguments:
         forecast (``xarray object``): Prediction ensemble with ``init`` dim renamed to
@@ -189,14 +170,12 @@ def _construct_init_lead_matrix(forecast, n, freq, leads):
             values corresponding to "real time", or ``init + lead`` over all inits and
             leads.
     """
-    # Note that `init` is renamed to `time` in compute functions.
-    dim = "init"  # was time
     init_lead_matrix = xr.concat(
         [
             xr.DataArray(
-                shift_cftime_index(forecast, dim, n, freq),
-                dims=[dim],
-                coords=[forecast[dim]],
+                shift_cftime_index(forecast, "init", n, freq),
+                dims=["init"],
+                coords=[forecast["init"]],
             )
             for n in n
         ],